refactor: directory structure for generated schema files
