+++ internal/generate/local_resources/local_group_resource_gen.go
