+++ internal/provider/local/local_group_data_source.go
@@ -3,7 +3,7 @@ package local
 import (
 	"context"
 	"fmt"
-	"terraform-provider-windows/internal/generator/local_datasources"
+	"terraform-provider-windows/internal/generate/local_datasources"
 
 	"github.com/d-strobel/gowindows"
 	"github.com/d-strobel/gowindows/windows/local"
