+++ internal/provider/local/local_user_resource.go
@@ -3,7 +3,7 @@ package local
 import (
 	"context"
 	"fmt"
-	"terraform-provider-windows/internal/generator/local_resources"
+	"terraform-provider-windows/internal/generate/local_resources"
 	"time"
 
 	"github.com/d-strobel/gowindows"
