+++ internal/schema/local_resources.json
