+++ internal/schema/provider_windows.json
